refactor(vm): split vm_var_ops.rs into cohesive submodules (§7-8)
`vm_var_ops.rs` (872 lines) exceeded the 500-line module convention. Split
into three cohesive files by responsibility, byte-identically (the function
multiset is verified unchanged; every method keeps its `pub(super)` visibility
and is reached via `self`, so behavior is preserved):

- `vm_var_ops.rs` (426): variable/hash/array entry resolution, Proxy closure
  env update, instance `:exists`, typed-container defaults, anon-state, Failure
  helpers, and the `native_element_default` free fn.
- `vm_var_multidim_helpers.rs` (187): multi-dimensional array depth, index
  coercion, and index/assign/delete helpers (the helper layer beneath the
  opcode handlers in `vm_var_multidim_ops.rs`).
- `vm_var_index_tracking.rs` (272): bound-index / element-share / deleted-index
  / initialized-index side-table bookkeeping.

Names align with the existing `vm_var_multidim_ops` / `vm_var_index_ops`
siblings. All three files are now under 500 lines.
